+++ nautobot/core/models/name_color_content_types.py
@@ -1,10 +1,12 @@
 from django.contrib.contenttypes.models import ContentType
 from django.db import models
+from django.utils.html import format_html
 
 from nautobot.core.choices import ColorChoices
 from nautobot.core.constants import CHARFIELD_MAX_LENGTH
 from nautobot.core.models import BaseManager, BaseModel, ContentTypeRelatedQuerySet
 from nautobot.core.models.fields import ColorField
+from nautobot.core.templatetags import helpers
 from nautobot.extras.models.change_logging import ChangeLoggedModel
 
 # Importing CustomFieldModel, ChangeLoggedModel, RelationshipModel from  nautobot.extras.models
@@ -54,3 +56,10 @@ class NameColorContentTypesModel(
 
     def get_content_types(self):
         return ",".join(f"{ct.app_label}.{ct.model}" for ct in self.content_types.all())
+
+    def get_color_display(self):
+        if self.color:
+            return format_html(
+                '<span class="label color-block" style="background-color: #{}">&nbsp;</span>', self.color
+            )
+        return helpers.placeholder(self.color)
